backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "StudyBuddy Agentic RAG backend"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "mangum>=0.17.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-pinecone>=0.2.0",
    "pinecone>=5.0.0",
    "langgraph>=0.2.0",
    "opik>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.9",
    "requests>=2.32.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rag"]
